afxdp/cos/queue_ops/v_min: relocate inline tests to v_min_tests.rs
v_min.rs is a flat file inside queue_ops/. 1,036-LOC inline test block extracted to sibling v_min_tests.rs (1,039 LOC incl 5-line header). Stub: `#[cfg(test)] #[path = "v_min_tests.rs"] mod tests;`.

After: v_min.rs drops from 1,217 → 184 LOC.

cargo build --release: clean. cargo test --release: 878 passed.
